% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_generators.R
\name{st_random_points}
\alias{st_random_points}
\title{Generate random points using Simple Sequential Inhibition}
\usage{
st_random_points(polygon, min_dist, n = Inf, planar_crs = NA, ...)
}
\arguments{
\item{polygon}{Boundary polygon}

\item{min_dist}{Minimum distance between points (in units of \code{polygon}'s CRS).}

\item{n}{How many points to place (defaults to \code{Inf}, meaning as many points as
possible will be fit)}

\item{planar_crs}{Planar CRS to use for coercion to \code{spatstat}-object; defaults
to Web Mercator.}

\item{...}{Additional arguments to pass to \code{spatstat.random::rSSI()}}
}
\value{
A \code{sf}-object of type POINT.
}
\description{
Wrapper for \code{spatstat.random::rSSI()} to generate quasi-random points within an \code{sf}-
POLYGON or MULTIPOLYGON (similar to \code{sf::st_sample()}). For some MULTIPOLYGONS,
you may need to turn spherical geometry off (\code{sf::sf_use_s2(FALSE)})
}
\examples{
library(ggplot2)

germany <- rnaturalearth::ne_countries(country = "Germany", returnclass = "sf")["admin"]
germany_pts <- st_random_points(germany, min_dist = 40000)

germany |> 
  ggplot() +
  geom_sf() +
  geom_sf(germany_pts, mapping = aes()) +
  theme_void()
}
