% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_generators.R
\name{st_random_voronoi}
\alias{st_random_voronoi}
\title{Random voronoi-tiles within polygon}
\usage{
st_random_voronoi(polygon, min_dist, n = Inf, planar_crs = NA, ...)
}
\arguments{
\item{polygon}{Boundary polygon (a POLYGON or MULTIPOLYGON)}

\item{min_dist}{Minimum distance between tiles' centroids (in units of \code{polygon}'s CRS)}

\item{n}{Number of tiles to generate (defaults to \code{Inf}, meaning as many as can be
fit without violating \code{minimum_distance} will be fit).}

\item{planar_crs}{Planar projection to use for generating centroids; defaults to
Web Mercator.}

\item{...}{Additional arguments to pass to \code{spatstat.random::rSSI()}, which handles the
quasi-random generation of centroids for the tiles.}
}
\value{
A \code{sf}-object of type MULTIPOLYGON.
}
\description{
Generate quasi-random voronoi-tiles within a \code{sf} POLYGON or MULTIPOLYGON. For
some MULTIPOLYGONS, you may need to turn spherical geometry off
(\code{sf::sf_use_s2(FALSE)})
}
\examples{
library(ggplot2)

germany <- rnaturalearth::ne_countries(country = "Germany", returnclass = "sf")["admin"]

germany_vor <- st_random_voronoi(germany, min_dist = 40000)
ggplot(germany_vor) +
  geom_sf() +
  theme_void()
}
